+++ actions/get_files.php
@@ -22,7 +22,7 @@
 	MongoConnection::connect();
 	MongoConnection::grid_fs();
 	$results = MongoConnection::grid_fs_find();
-	$results->sort(array("uploadDate"));
+	$results->sort(array("uploadDate" => -1));
 	MongoConnection::close();
 	
 	foreach($results as $result) {
@@ -39,8 +39,8 @@
 		
 		foreach($files as $file) {
 			$html .= '<tr id="' . $file['_id'] . '" class="file"><td><input type="checkbox" class="file-delete-check" value="' . $file['_id'] . '" /></td><td><a href="' . Links::render("download-file", array($file['_id']->__toString())) . '" class="btn btn-mini"><i class="icon-download-alt"></i></a>';
-			$html .= (strpos($file['type'], 'text') !== false) ? ' <a href="' . Links::render("view-file-raw", array($file['_id']->__toString())) . '" class="btn btn-mini"><i class="icon-align-left"></i>' : null;
-			$html .= '</td><td><a class="btn btn-mini disabled expand-west expand-file-id">' . Functions::add_ellipsis_reverse($file['_id'], 4) . '</a></td><td><a href="' . Links::render("view-file", array($file['_id']->__toString())) . '">' . $file['real_filename'] . '</a></td><td><span class="badge badge-info">';
+			$html .= (strpos($file['type'], 'text') !== false || $file['type'] === "application/json") ? ' <a href="' . Links::render("view-file-raw", array($file['_id']->__toString())) . '" class="btn btn-mini"><i class="icon-align-left"></i>' : null;
+			$html .= '</td><td><a class="btn btn-mini disabled expand-west expand-file-id">' . Functions::add_ellipsis_reverse($file['_id'], 8) . '</a></td><td><a href="' . Links::render("view-file", array($file['_id']->__toString())) . '">' . $file['real_filename'] . '</a></td><td><span class="badge badge-info">';
 			$html .= empty($file['type']) ? "unknown" : $file['type'];
 			$html .= '</span></td><td>' . Functions::format_bytes($file['length']) . '</td><td>' . $file['uploadDate'] . '</td></tr>';
 		}